backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
asyncpg==0.29.0
alembic==1.13.1
psutil==5.9.8
requests==2.31.0
prometheus-client==0.19.0
python-dateutil==2.8.2
apscheduler==3.10.4